+++ buildSrc/src/main/java/MyApp.kt
@@ -2,6 +2,6 @@
 
 object MyApp {
     const val appId = "com.sysadmindoc.iosicons"
-    const val version = 5
-    const val versionName = "1.1.3"
+    const val version = 6
+    const val versionName = "1.1.4"
 }